refactor(ofs): Split fuse3 impl into fuse3_opendal
