+++ integrations/fuse3/Cargo.toml
@@ -16,11 +16,8 @@
 # under the License.
 
 [package]
-categories = ["filesystem"]
-description = "OpenDAL File System"
-keywords = ["storage", "data", "s3", "fs", "azblob"]
-name = "ofs"
-version = "0.0.4"
+description = "fuse3 integration for Apache OpenDAL"
+name = "fuse3_opendal"
 
 authors = ["Apache OpenDAL <dev@opendal.apache.org>"]
 edition = "2021"
@@ -28,39 +25,16 @@ homepage = "https://opendal.apache.org/"
 license = "Apache-2.0"
 repository = "https://github.com/apache/opendal"
 rust-version = "1.75"
+version = "0.1.0"
 
 [dependencies]
-anyhow = "1"
-bytes = "1.5.0"
-chrono = "0.4.34"
-clap = { version = "4.5.1", features = ["derive", "env"] }
-env_logger = "0.11.2"
-fuse3 = { "version" = "0.7.2", "features" = ["tokio-runtime", "unprivileged"] }
-futures-util = "0.3.30"
-libc = "0.2.154"
-log = "0.4.21"
-nix = { version = "0.28.0", features = ["user"] }
 opendal = { version = "0.47.0", path = "../../core" }
-fuse3_opendal = { version = "0.1.0", path = "../../integrations/fuse3" }
+fuse3 = { version = "0.7.2", "features" = ["tokio-runtime", "unprivileged"] }
+bytes = "1.6.0"
+futures-util = "0.3.30"
 sharded-slab = "0.1.7"
-tokio = { version = "1.37.0", features = [
-    "fs",
-    "macros",
-    "rt-multi-thread",
-    "io-std",
-    "signal",
-] }
-url = "2.5.0"
-
-[features]
-default = ["services-fs", "services-s3"]
-services-fs = ["opendal/services-fs"]
-services-s3 = ["opendal/services-s3"]
+tokio = "1.38.0"
+libc = "0.2.155"
+log = "0.4.21"
 
 [dev-dependencies]
-tempfile = "3.10.1"
-test-context = "0.3.0"
-urlencoding = "2.1.3"
-uuid = "1.7.0"
-walkdir = "2.5.0"
-opendal = { version = "0.47.0", path = "../../core", features = ["tests"] }
